refactor(api): change how does the private route and works
